Moving keyboard and mouse to internal.
