+++ container/focus_test.go
@@ -24,7 +24,7 @@ import (
 	"github.com/mum4k/termdash/internal/draw"
 	"github.com/mum4k/termdash/internal/event"
 	"github.com/mum4k/termdash/internal/event/testevent"
-	"github.com/mum4k/termdash/mouse"
+	"github.com/mum4k/termdash/internal/mouse"
 	"github.com/mum4k/termdash/terminal/faketerm"
 	"github.com/mum4k/termdash/terminalapi"
 )
