+++ internal/mouse/button/button.go
@@ -18,7 +18,7 @@ package button
 import (
 	"image"
 
-	"github.com/mum4k/termdash/mouse"
+	"github.com/mum4k/termdash/internal/mouse"
 	"github.com/mum4k/termdash/terminalapi"
 )
