+++ internal/mouse/button/button_test.go
@@ -19,7 +19,7 @@ import (
 	"image"
 	"testing"
 
-	"github.com/mum4k/termdash/mouse"
+	"github.com/mum4k/termdash/internal/mouse"
 	"github.com/mum4k/termdash/terminalapi"
 )
