+++ internal/mouse/mouse.go
