+++ terminal/termbox/event_test.go
@@ -21,8 +21,8 @@ import (
 	"testing"
 
 	"github.com/kylelemons/godebug/pretty"
-	"github.com/mum4k/termdash/keyboard"
-	"github.com/mum4k/termdash/mouse"
+	"github.com/mum4k/termdash/internal/keyboard"
+	"github.com/mum4k/termdash/internal/mouse"
 	"github.com/mum4k/termdash/terminalapi"
 
 	tbx "github.com/nsf/termbox-go"
