+++ widgets/button/options.go
@@ -22,7 +22,7 @@ import (
 
 	"github.com/mum4k/termdash/internal/cell"
 	"github.com/mum4k/termdash/internal/cell/runewidth"
-	"github.com/mum4k/termdash/keyboard"
+	"github.com/mum4k/termdash/internal/keyboard"
 	"github.com/mum4k/termdash/widgetapi"
 )
